made the 'outside' worker case the default
